feat: add learn-page component
