+++ src/components/english-puzzle/components/controller/game.controller.js
@@ -132,7 +132,6 @@ class GameController {
         new ErrorPopup().openPopup({ text: error.message });
         return null;
       });
-    debugger;
     if (!this.fetchedRoundData) return;
     const sentences = this.fetchedRoundData.map((wordData) => wordData.textExample);
