+++ src/components/main/common/main.constants.js
@@ -1,9 +1,3 @@
-const MAIN_GREETINGS = 'Hello from Main';
-const GAME_BLOCK = document.querySelector('.main');
-const TEMPLATE_MAIN_GAMEINTRO = document.getElementById('template-main__game-intro');
-const TEMPLATE_MAIN_GAME = document.getElementById('template-main__game');
-const INPUT_AREA = document.querySelector('body > div.main__game > div.card-container > div.card > div.card-main > div.learn-content > div > div.input-container > input');
+const DEFAULT_ENGLISH_LEVEL = 0;
 
-export {
-  MAIN_GREETINGS, GAME_BLOCK, TEMPLATE_MAIN_GAMEINTRO, TEMPLATE_MAIN_GAME, INPUT_AREA,
-};
+export { DEFAULT_ENGLISH_LEVEL };
