--- src/components/main/components/mainGame/data.js
@@ -1,170 +1,0 @@
-export const defaultData = [{
-  audio: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0011.mp3',
-  audioExample: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0011_example.mp3',
-  audioMeaning: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0011_meaning.mp3',
-  group: 0,
-  id: '5e9f5ee35eb9e72bc21af4aa',
-  image: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0011.jpg',
-  page: 0,
-  textExample: 'The woman <b>enjoys</b> riding her bicycle.',
-  textExampleTranslate: 'Женщина любит кататься на велосипеде',
-  textMeaning: 'To <i>enjoy</i> is to like something.',
-  textMeaningTranslate: 'Наслаждаться значит любить что-то',
-  transcription: '[indʒɔ́i]',
-  word: 'enjoy',
-  wordTranslate: 'наслаждаться',
-  wordsPerExampleSentence: 6,
-},
-{
-  audio: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0012.mp3',
-  audioExample: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0012_example.mp3',
-  audioMeaning: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0012_meaning.mp3',
-  group: 0,
-  id: '5e9f5ee35eb9e72bc21af4ab',
-  image: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0012.jpg',
-  page: 0,
-  textExample: 'I will <b>invite</b> my friends to my birthday party.',
-  textExampleTranslate: 'Я приглашаю своих друзей на мой день рождения',
-  textMeaning: 'To <i>invite</i> is to ask someone to come to a place or event.',
-  textMeaningTranslate: 'Пригласить - это попросить кого-нибудь прийти на место или событие',
-  transcription: '[inváit]',
-  word: 'invite',
-  wordTranslate: 'пригласить',
-  wordsPerExampleSentence: 9,
-},
-{
-  audio: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0014.mp3',
-  audioExample: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0014_example.mp3',
-  audioMeaning: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0014_meaning.mp3',
-  group: 0,
-  id: '5e9f5ee35eb9e72bc21af4ac',
-  image: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0014.jpg',
-  page: 0,
-  textExample: 'January is the first <b>month</b> of the year.',
-  textExampleTranslate: 'январь - первый месяц года',
-  textMeaning: 'A <i>month</i> is one of 12 periods of time in one year.',
-  textMeaningTranslate: 'Месяц - это один из 12 периодов времени в году',
-  transcription: '[mʌnθ]',
-  word: 'month',
-  wordTranslate: 'месяц',
-  wordsPerExampleSentence: 8,
-},
-{
-  audio: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0015.mp3',
-  audioExample: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0015_example.mp3',
-  audioMeaning: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0015_meaning.mp3',
-  group: 0,
-  id: '5e9f5ee35eb9e72bc21af4ad',
-  image: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0015.jpg',
-  page: 0,
-  textExample: 'They will <b>travel</b> to Argentina this summer.',
-  textExampleTranslate: 'Этим летом они отправятся в Аргентину',
-  textMeaning: 'To <i>travel</i> is to go to a faraway place on vacation or business.',
-  textMeaningTranslate: 'Путешествовать - это отправиться в далекое место на отдых или по делам',
-  transcription: '[trǽvəl]',
-  word: 'travel',
-  wordTranslate: 'путешествовать',
-  wordsPerExampleSentence: 7,
-},
-{
-  audio: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0013.mp3',
-  audioExample: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0013_example.mp3',
-  audioMeaning: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0013_meaning.mp3',
-  group: 0,
-  id: '5e9f5ee35eb9e72bc21af4ae',
-  image: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0013.jpg',
-  page: 0,
-  textExample: 'I <b>love</b> my family very much.',
-  textExampleTranslate: 'Я очень люблю свою семью',
-  textMeaning: 'To <i>love</i> is to like something or someone a lot.',
-  textMeaningTranslate: 'Любить значит любить что-то или кого-то много',
-  transcription: '[lʌv]',
-  word: 'love',
-  wordTranslate: 'любовь',
-  wordsPerExampleSentence: 6,
-},
-{
-  audio: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0016.mp3',
-  audioExample: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0016_example.mp3',
-  audioMeaning: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0016_meaning.mp3',
-  group: 0,
-  id: '5e9f5ee35eb9e72bc21af4af',
-  image: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0016.jpg',
-  page: 0,
-  textExample: 'My <b>typical</b> breakfast is toast and eggs.',
-  textExampleTranslate: 'Мой типичный завтрак - тост и яйца',
-  textMeaning: 'If something is <i>typical</i>, it is normal, or something that usually happens.',
-  textMeaningTranslate: 'Если что-то типичное, это нормально, или что-то, что обычно происходит',
-  transcription: '[típikəl]',
-  word: 'typical',
-  wordTranslate: 'типичный',
-  wordsPerExampleSentence: 7,
-},
-{
-  audio: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0017.mp3',
-  audioExample: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0017_example.mp3',
-  audioMeaning: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0017_meaning.mp3',
-  group: 0,
-  id: '5e9f5ee35eb9e72bc21af4b0',
-  image: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0017.jpg',
-  page: 0,
-  textExample: 'She wants to <b>visit</b> her grandmother.',
-  textExampleTranslate: 'Она хочет навестить свою бабушку',
-  textMeaning: 'To <i>visit</i> is to go and spend time in another place or see another person.',
-  textMeaningTranslate: 'Посетить - значит пойти и провести время в другом месте или увидеть другого человека',
-  transcription: '[vízit]',
-  word: 'visit',
-  wordTranslate: 'посещение',
-  wordsPerExampleSentence: 6,
-},
-{
-  audio: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0018.mp3',
-  audioExample: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0018_example.mp3',
-  audioMeaning: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0018_meaning.mp3',
-  group: 0,
-  id: '5e9f5ee35eb9e72bc21af4b1',
-  image: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0018.jpg',
-  page: 0,
-  textExample: 'Today’s <b>weather</b> is rainy and cloudy.',
-  textExampleTranslate: 'Сегодня погода дождливая и облачная',
-  textMeaning: '<i>Weather</i> is the temperature and the state of the outdoors.',
-  textMeaningTranslate: 'Погода это температура и состояние на улице',
-  transcription: '[weðər]',
-  word: 'weather',
-  wordTranslate: 'погода',
-  wordsPerExampleSentence: 6,
-},
-{
-  audio: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0020.mp3',
-  audioExample: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0020_example.mp3',
-  audioMeaning: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0020_meaning.mp3',
-  group: 0,
-  id: '5e9f5ee35eb9e72bc21af4b2',
-  image: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/01_0020.jpg',
-  page: 0,
-  textExample: 'The store carried both red and white <b>wine</b>.',
-  textExampleTranslate: 'В магазине было красное и белое вино',
-  textMeaning: '<i>Wine</i> is an alcoholic drink made from grapes.',
-  textMeaningTranslate: 'Вино - это алкогольный напиток из винограда',
-  transcription: '[wain]',
-  word: 'wine',
-  wordTranslate: 'вино',
-  wordsPerExampleSentence: 8,
-},
-{
-  audio: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/02_0021.mp3',
-  audioExample: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/02_0021_example.mp3',
-  audioMeaning: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/02_0021_meaning.mp3',
-  group: 0,
-  id: '5e9f5ee35eb9e72bc21af4b4',
-  image: 'https://raw.githubusercontent.com/caspercarver/rslang-data/master/files/02_0021.jpg',
-  page: 1,
-  textExample: 'Riding in the rough water was an <b>adventure</b>.',
-  textExampleTranslate: 'Езда в бурной воде была приключением',
-  textMeaning: 'An <i>adventure</i> is a fun or exciting thing that you do.',
-  textMeaningTranslate: 'Приключение - это забавная или захватывающая вещь, которую ты делаешь',
-  transcription: '[ədvéntʃər]',
-  word: 'adventure',
-  wordTranslate: 'приключение',
-  wordsPerExampleSentence: 8,
-}];
