+++ src/components/promo/promo.index.html
@@ -23,11 +23,8 @@
               <div>
                 <button type="submit" class="login__submitBtn">
                   <img src="./assets/promo/img/icon/vector.svg" alt="logo">
-                  SIGN IN
+                  Let`s try!
                 </button>
-                <a class="login__submitBtn" href="main.index.html">
-                  Main Application
-                </a>
               </div>
             </div>
           </div>
